+++ halo2-base/Cargo.toml
@@ -23,6 +23,9 @@ halo2_proofs = { git = "https://github.com/privacy-scaling-explorations/halo2.gi
 plotters = { version = "0.3.0", optional = true }
 tabbycat = { version = "0.1", features = ["attributes"], optional = true }
 
+# test-utils
+rand = { version = "0.8", optional = true }
+
 [dev-dependencies]
 ark-std = { version = "0.3.0", features = ["print-trace"] }
 rand = "0.8"
@@ -44,6 +47,7 @@ halo2-pse = ["halo2_proofs"]
 halo2-axiom = ["halo2_proofs_axiom"]
 display = []
 profile = ["halo2_proofs_axiom?/profile"]
+test-utils = ["dep:rand"]
 
 [[bench]]
 name = "mul"
