+++ halo2-base/src/gates/mod.rs
@@ -2,7 +2,7 @@ pub mod builder;
 pub mod flex_gate;
 pub mod range;
 
-#[cfg(test)]
+#[cfg(any(test, feature = "test-utils"))]
 pub mod tests;
 
 pub use flex_gate::{GateChip, GateInstructions};
