+++ halo2-ecc/src/fields/tests/mod.rs
@@ -1,0 +1,2 @@
+pub mod fp;
+pub mod fp12;
